Require authentication for relevant endpoints

+++ server/api_server/api.py
@@ -265,7 +265,7 @@ class UsersAPI(Resource):
                 type: string
                 required: false
         security:
-
+            - bearerAuth: []
         responses:
             200:
                 description: The updated user
@@ -315,7 +315,7 @@ class UsersAPI(Resource):
                             default: []
         """
         if not user_is_logged_in(user_id):
-            return {'error' : 'login required'}, 401
+            return NOT_LOGGED_IN_RESPONSE
         parser = reqparse.RequestParser()
         for arg, datatype in User.constructor_params().items():
             parser.add_argument(arg, type=datatype)
@@ -552,6 +552,8 @@ class FriendsAPI(Resource):
                 in: body
                 type: integer
                 required: true
+        security:
+            - bearerAuth: []
         responses:
             200:
                 description: The resulting friend-pair
@@ -565,6 +567,8 @@ class FriendsAPI(Resource):
                             type: integer
                             description: An echo of {friendId}
         """
+        if not user_is_logged_in(user_id):
+            return NOT_LOGGED_IN_RESPONSE
         args = self._parse_put_delete()
         # Create symmetrical entities
         friends = Friends(user_1=user_id, user_2=args['friendId'])
@@ -591,6 +595,8 @@ class FriendsAPI(Resource):
                 in: body
                 type: integer
                 required: true
+        security:
+            - bearerAuth: []
         responses:
             200:
                 description: The deleted friend-pair
@@ -604,6 +610,8 @@ class FriendsAPI(Resource):
                             type: integer
                             description: An echo of {friendId}
         """
+        if not user_is_logged_in(user_id):
+            return NOT_LOGGED_IN_RESPONSE
         args = self._parse_put_delete()
         friends = Friends.query.filter(
             Friends.user_1 == user_id, Friends.user_2 == args['friendId']
@@ -777,6 +785,8 @@ class DraftsAPI(Resource):
                 in: body
                 required: true
                 description: The unique player ID of the player to draft
+        security:
+            - bearerAuth: []
         responses:
             200:
                 description: The created draft entity
@@ -794,6 +804,8 @@ class DraftsAPI(Resource):
                                 of errors: the user's draft being full and 
                                 integrity errors due to the passed parameters.
         """
+        if not user_is_logged_in(user_id):
+            return NOT_LOGGED_IN_RESPONSE
         parser = reqparse.RequestParser()
         parser.add_argument('playerId', type=int)
         args = parser.parse_args(strict=True)
@@ -839,11 +851,15 @@ class DraftsAPI(Resource):
                 in: body
                 required: true
                 description: The unique player ID of the player to draft
+        security:
+            - bearerAuth: []
         responses:
             200:
                 description: The removed draft entity
                 schema:
         """
+        if not user_is_logged_in(user_id):
+            return NOT_LOGGED_IN_RESPONSE
         parser = reqparse.RequestParser()
         parser.add_argument('playerId', type=int)
         args = parser.parse_args(strict=True)
@@ -926,6 +942,8 @@ class LeagueAPI(Resource):
                 type: integer
                 description: The ID of the league to delete
                 required: true
+        security:
+            - bearerAuth: []
         responses:
             200:
                 description: The deleted entity
@@ -934,6 +952,8 @@ class LeagueAPI(Resource):
         """
         league = FantasyLeague.query.filter(
             FantasyLeague.league_id == league_id).first()
+        if not user_is_logged_in(league.owner_id):
+            return NOT_LOGGED_IN_RESPONSE
         db.session.delete(league)
         db.session.commit()
         return fantasy_league_schema.jsonify(league)
@@ -969,6 +989,8 @@ class LeagueAPI(Resource):
                 required: true
                 type: string
                 description: The name of the league
+        security:
+            - bearerAuth: []
         responses:
             200:
                 description: The created fantasy league
@@ -982,6 +1004,8 @@ class LeagueAPI(Resource):
         parser.add_argument('public', type=bool)
         parser.add_argument('name', type=str)
         args = parser.parse_args(strict=True)
+        if not user_is_logged_in(args['ownerId']):
+            return NOT_LOGGED_IN_RESPONSE
         league = FantasyLeague(event_id=args['eventId'],
                                owner_id=args['ownerId'],
                                draft_size=args['draftSize'],
@@ -1018,6 +1042,8 @@ class LeagueAPI(Resource):
                 required: false
                 type: string
                 description: The name of the league
+        security:
+            - bearerAuth: []
         responses:
             200:
                 description: The updated fantasy league
@@ -1032,6 +1058,8 @@ class LeagueAPI(Resource):
         league = FantasyLeague.query.filter(FantasyLeague.league_id
                                             == league_id).first()
         if league is not None:
+            if not user_is_logged_in(league.owner_id):
+                return NOT_LOGGED_IN_RESPONSE
             args = parser.parse_args(strict=True)
             for key, value in args.items():
                 if value is not None:
@@ -1139,7 +1167,7 @@ def user_is_logged_in(user_id):
         scheme, token = header.split(' ')
     except ValueError:
         return False
-    if scheme != 'bearer':
+    if scheme.lower() != 'bearer':
         return False
     user = User.query.filter(User.user_id == user_id).first()
     email, hashed = base64.b64decode(token).decode().split(':')
@@ -1158,6 +1186,7 @@ api.add_resource(LeagueAPI, '/leagues/<int:league_id>')
 api.add_resource(EntrantsAPI, '/entrants/<int:event_id>')
 api.add_resource(LoginAPI, '/login')
 
+NOT_LOGGED_IN_RESPONSE = [{'error' : 'login required'}, 401]
 
 def make_pagination_reqparser():
     parser = reqparse.RequestParser(bundle_errors=True)
